requirements.txt
Django==1.10.2
django-admin-bootstrapped==2.5.7
django-debug-toolbar==1.6
django-registration-redux==1.4
django-showurls==1.0.1
djangorestframework==3.5.1
dj-database-url==0.4.1
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2.2

